fix previous unfinished merge conflicts

+++ src/noggit/MapTile.cpp
@@ -1614,27 +1614,27 @@ void MapTile::setVertexColorImage(QImage const& baseimage, int mode, bool tiledE
             case 1: // Add
             {
               auto color = image.pixelColor((k * 16) + x, (l * 16) + y);
-              colors[idx].x =  std::min(2.0f, std::max(0.0f, colors[idx].x + color.redF() * 2.f));
-              colors[idx].y =  std::min(2.0f, std::max(0.0f, colors[idx].y + color.greenF() * 2.f));
-              colors[idx].z =  std::min(2.0f, std::max(0.0f, colors[idx].z + color.blueF() * 2.f));
+              colors[idx].x = std::min(2.0f, std::max(0.0f, colors[idx].x + static_cast<float>(color.redF() * 2.f)));
+              colors[idx].y =  std::min(2.0f, std::max(0.0f, colors[idx].y + static_cast<float>(color.greenF() * 2.f)));
+              colors[idx].z =  std::min(2.0f, std::max(0.0f, colors[idx].z + static_cast<float>(color.blueF() * 2.f)));
               break;
             }
 
             case 2: // Subtract
             {
               auto color = image.pixelColor((k * 16) + x, (l * 16) + y);
-              colors[idx].x =  std::min(2.0f, std::max(0.0f, colors[idx].x - color.redF() * 2.f));
-              colors[idx].y =  std::min(2.0f, std::max(0.0f, colors[idx].y - color.greenF() * 2.f));
-              colors[idx].z =  std::min(2.0f, std::max(0.0f, colors[idx].z - color.blueF() * 2.f));
+              colors[idx].x =  std::min(2.0f, std::max(0.0f, colors[idx].x - static_cast<float>(color.redF() * 2.f)));
+              colors[idx].y =  std::min(2.0f, std::max(0.0f, colors[idx].y - static_cast<float>(color.greenF() * 2.f)));
+              colors[idx].z =  std::min(2.0f, std::max(0.0f, colors[idx].z - static_cast<float>(color.blueF() * 2.f)));
               break;
             }
 
             case 3: // Multiply
             {
               auto color = image.pixelColor((k * 16) + x, (l * 16) + y);
-              colors[idx].x =  std::min(2.0f, std::max(0.0f, colors[idx].x * color.redF() * 2.f));
-              colors[idx].y =  std::min(2.0f, std::max(0.0f, colors[idx].y * color.greenF() * 2.f));
-              colors[idx].z =  std::min(2.0f, std::max(0.0f, colors[idx].z * color.blueF() * 2.f));
+              colors[idx].x =  std::min(2.0f, std::max(0.0f, colors[idx].x * static_cast<float>(color.redF() * 2.f)));
+              colors[idx].y =  std::min(2.0f, std::max(0.0f, colors[idx].y * static_cast<float>(color.greenF() * 2.f)));
+              colors[idx].z =  std::min(2.0f, std::max(0.0f, colors[idx].z * static_cast<float>(color.blueF() * 2.f)));
               break;
             }
           }
